Feat: Rubocop prefers nested modules to compact
While 'compact' definitely looks nicer and results in a shorter
tabbed-in amount, it does change the behavior. Constant lookup is
changed so that only the fully qualified constant is added to ancestors
and the intermediary steps are not. This is usually not an issue.

The big problem is that ruby will throw an error if the intermediary
modules aren't defined before they're used in the compact definition.
This causes issues when loading files and makes the order important,
which can cause it to work in tests and fail in development or
vice-versa.