PowerShell: Fix casing issue w/ grab-svn-authors
There was a bug in the script that surfaced while trying to use the
transformed authors txt. Some lines starting with 'jdoe' in the were
transformed to 'JDoe' in the output, based on the casing in their
official AD account SamAccountName property. The change is to use the
samAccountName variable we pulled from the script, not from the AD
account!