Zsh: Fix git_pretty_log wrapping and color
Fixes a few things:

- Wrapping on Windows is looking good. This might introduce a wrapping
problem on macOS, but I'm not using that as much recently

- The refs portion is now correctly highlighted yellow. Previously it
wasn't working when the refs were on a new line. On Windows and macOS.
Something is bugged with the '%d' I had previously. Adding a trailing
space and changing the refs to '%D', which removes the parens, seems
to fix it.